src/requirements.txt
aiohttp
pillow
py-cord